Add more docs, reorganize stable stuff
