+++ pyro/infer/reparam/__init__.py
@@ -1,7 +1,6 @@
-from .levy_ito import StableHMMReparam
 from .loc_scale import LocScaleReparam
 from .neutra import NeuTraReparam
-from .stable import StableReparam, SymmetricStableReparam
+from .stable import StableHMMReparam, StableReparam, SymmetricStableReparam
 from .transform import TransformReparam
 
 __all__ = [
